BitcoinNodeWalletProvider: Move wallet related functions to own provider
