+++ packages/bitcoin-node-wallet-provider/lib/index.js
@@ -1,0 +1,3 @@
+import BitcoinNodeWalletProvider from './BitcoinNodeWalletProvider'
+
+export default BitcoinNodeWalletProvider
