+++ packages/bitcoin-node-wallet-provider/package.json
@@ -1,0 +1,44 @@
+{
+    "name": "@liquality/bitcoin-node-wallet-provider",
+    "umdName": "BitcoinNodeWalletProvider",
+    "umdExport": "default",
+    "version": "0.2.11",
+    "description": "",
+    "module": "lib/index.js",
+    "main": "dist/index.cjs.js",
+    "files": [
+      "dist",
+      "lib"
+    ],
+    "scripts": {
+      "build:node": "webpack --config ../../webpack/webpack.node.config.js",
+      "build:browser": "webpack --config ../../webpack/webpack.browser.config.js",
+      "build": "webpack --config ../../webpack/webpack.config.js"
+    },
+    "author": "Liquality <info@liquality.io>",
+    "license": "MIT",
+    "dependencies": {
+      "@babel/runtime": "^7.4.3",
+      "@liquality/jsonrpc-provider": "^0.2.11",
+      "@liquality/wallet-provider": "^0.2.10",
+      "@liquality/utils": "^0.2.11",
+      "bignumber.js": "^8.1.1",
+      "lodash": "^4.17.11"
+    },
+    "engines": {
+      "node": "~8.12.0"
+    },
+    "homepage": "https://github.com/liquality/chainabstractionlayer#readme",
+    "repository": {
+      "type": "git",
+      "url": "git+https://github.com/liquality/chainabstractionlayer.git"
+    },
+    "bugs": {
+      "url": "https://github.com/liquality/chainabstractionlayer/issues"
+    },
+    "publishConfig": {
+      "access": "public"
+    },
+    "gitHead": "1855d9ec369c8c577f74eb04e310a7c00b9eaa6c"
+  }
+
