+++ packages/bundle/package.json
@@ -24,6 +24,7 @@
     "@liquality/bitcoin-ledger-provider": "^0.2.11",
     "@liquality/bitcoin-networks": "^0.2.11",
     "@liquality/bitcoin-rpc-provider": "^0.2.11",
+    "@liquality/bitcoin-node-wallet-provider": "^0.2.11",
     "@liquality/bitcoin-swap-provider": "^0.2.11",
     "@liquality/bitcoin-utils": "^0.2.11",
     "@liquality/client": "^0.2.11",
